SCRUM-4_Implements_JWT-token_(security-config)

+++ src/main/java/de/telran/onlineshopgarden/service/UserService.java
@@ -4,14 +4,17 @@ import de.telran.onlineshopgarden.dto.UserCreateDto;
 import de.telran.onlineshopgarden.dto.UserDto;
 import de.telran.onlineshopgarden.dto.UserUpdateDto;
 import de.telran.onlineshopgarden.entity.User;
+import de.telran.onlineshopgarden.exception.AccessForbiddenException;
 import de.telran.onlineshopgarden.exception.ResourceNotFoundException;
 import de.telran.onlineshopgarden.mapper.UserMapper;
 import de.telran.onlineshopgarden.repository.CartItemRepository;
 import de.telran.onlineshopgarden.repository.CartRepository;
 import de.telran.onlineshopgarden.repository.FavoriteRepository;
 import de.telran.onlineshopgarden.repository.UserRepository;
+import de.telran.onlineshopgarden.security.AuthService;
 import jakarta.persistence.EntityManager;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,9 +31,11 @@ public class UserService {
     private final CartRepository cartRepository;
     private final CartItemRepository cartItemRepository;
     private final EntityManager entityManager;
+    private final PasswordEncoder passwordEncoder;
+    private final AuthService authService;
 
     @Autowired
-    public UserService(UserRepository repository, UserMapper mapper, OrderService orderService, FavoriteRepository favoriteRepository, CartRepository cartRepository, CartItemRepository cartItemRepository, EntityManager entityManager) {
+    public UserService(UserRepository repository, UserMapper mapper, OrderService orderService, FavoriteRepository favoriteRepository, CartRepository cartRepository, CartItemRepository cartItemRepository, EntityManager entityManager, PasswordEncoder passwordEncoder, AuthService authService) {
         this.repository = repository;
         this.mapper = mapper;
         this.orderService = orderService;
@@ -38,6 +43,8 @@ public class UserService {
         this.cartRepository = cartRepository;
         this.cartItemRepository = cartItemRepository;
         this.entityManager = entityManager;
+        this.passwordEncoder = passwordEncoder;
+        this.authService = authService;
     }
 
     public List<UserDto> getAll() {
@@ -53,23 +60,38 @@ public class UserService {
     @Transactional
     public UserDto create(UserCreateDto dto) {
         User user = mapper.createDtoToEntity(dto);
-        // TODO: add password encryption
+        user.setPasswordHash(passwordEncoder.encode(dto.getPassword()));
         return mapper.entityToDto(repository.save(user));
     }
 
    @Transactional
     public UserDto update(Integer id, UserUpdateDto dto) {
-        User user = repository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
-        user.setName(dto.getName());
+       String login = authService.getAuthInfo().getLogin();
+       User user = repository.findUserByEmail(login).get();
+
+       repository.findById(id)
+               .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
+
+       if (!id.equals(user.getUserId())) {
+           throw new AccessForbiddenException("You cannot update another user");
+       }
+
+       user.setName(dto.getName());
         user.setPhoneNumber(dto.getPhone());
         return mapper.entityToDto(user);
     }
 
     @Transactional
     public void mask(Integer id) {
-        User user = repository.findById(id)
+        String login = authService.getAuthInfo().getLogin();
+        User user = repository.findUserByEmail(login).get();
+
+        repository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("User with id " + id + " not found"));
+
+        if (!id.equals(user.getUserId())) {
+            throw new AccessForbiddenException("You cannot delete another user");
+        }
 
         favoriteRepository.deleteByUserUserId(id);
 
@@ -87,4 +109,5 @@ public class UserService {
         user.setPhoneNumber("0000000000");
         user.setPasswordHash("DELETED");
     }
+
 }
